fix(nextjs): Fix serializable issue when build with tsup & tsup config cleanup
Build was setting NODE_ENV to production and using it to
determine minification and some other building variables.
By ovveriding the NODE_ENV in tsup command run, the env
of the whole package was being set to production even when
it was being used in development which caused the fix for
serializable props to be ignored and the code to fail.
To fix this we omitted the NODE_ENV ovveride and used
`--watch` option to determine minification on build.